Refactor logic into ValidationHoc
