+++ magda-web-client/src/Components/Dataset/Add/DatasetAddCommon.scss
@@ -84,7 +84,7 @@
         border: solid 1px #979797;
     }
 
-    .select-box {
+    .react-select-with-validation-container {
         .au-error-text {
             font-size: 16px;
             font-size: 1rem;
